+++ tests/test_model_v2_EdmStructTypeSerializer.py
@@ -9,7 +9,7 @@ from pyodata.exceptions import PyODataException
 @pytest.fixture
 def complex_type_property_declarations():
     return {
-        'TestString': (Types.parse_type_name('Edm.String'), "'FooBar'", "'FooBar'", 'FooBar'),
+        'TestString': (Types.parse_type_name('Edm.String'), 'FooBar', "'FooBar'", 'FooBar'),
         'TestBoolean': (Types.parse_type_name('Edm.Boolean'), False, 'false', False),
         'TestInt64': (Types.parse_type_name('Edm.Int64'), '123L', '123L', 123),
         'TestDateTime': (Types.parse_type_name('Edm.DateTime'), "/Date(2147483647000)/", "datetime'2038-01-19T03:14:07'",